refactor(bin-designer): split PreviewCanvas into focused modules
Decompose the 797-LOC PreviewCanvas.tsx into 2 sibling files. The
component file is 407 LOC focused on canvas wiring + state; camera
infrastructure and overlay components live in siblings.

Extracted modules:

- previewCanvasCamera.tsx (296) — camera infrastructure:
  `calculateIdealDistance` + `calculateBinCenter` (FOV math),
  `CAMERA_PRESETS` map, `CameraController` component (must run inside
  Canvas to access useThree/useFrame), `usePresetTransition` hook
  (spherical-coord interpolation between presets), and `SceneLighting`
  (theme-aware 3-point lights). The non-component exports
  (CAMERA_PRESETS, usePresetTransition) carry a file-level
  react-refresh/only-export-components disable since they're
  co-located with their paired components by design.
- previewCanvasOverlays.tsx (128) — overlay components rendered
  outside the Canvas: `TouchHint` (first-visit mobile gesture
  reminder, dismissed via settings.dismissedHints) and
  `GeneratingIndicator` (cycling SimCity-style loading messages
  with reduced-motion support).

PreviewCanvas.tsx: 797 LOC -> 407 LOC. The component file now contains
only the orchestration: ref/state setup, store reads, color persistence,
keyboard shortcuts wiring, double-tap reset, and the JSX tree.

1 test file / 13 tests pass (1.5s targeted run).
